+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the Solarium library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Solarium\Component\Result\Grouping\ValueGroup::getValue() returns `string|float|int|bool|null` according to field type or function return type instead of always coercing to `?string`
+
+
 ## [7.0.0]
 ### Added
 - Solr 10 support